refactor: split UserGuideDialog and validation.ts for coding standards compliance
- Split UserGuideDialog.tsx (558→101 lines) by extracting static content to
  src/components/userGuide/guideSteps.tsx (467 lines)
- Split validation.ts (598→38 lines) into focused modules:
  - validation/types.ts: Type definitions
  - validation/schemas.ts: Entity schemas
  - validation/schemaValidator.ts: Core validation logic
  - validation/sanitization.ts: Data sanitization
  - validation/entityValidators.ts: Entity-specific validators
  - validation/relationshipValidator.ts: Data relationship checks
- Removed stale backup file TouchDebugInfo.tsx.backup
- All original imports preserved via re-export pattern
